// apps/books/src/pages/reports/reportTabs/balancesheet/types.ts
export interface FinancialData {
  category: string
  isHeader?: boolean
  isTotal?: boolean
  isTotalEquity?: boolean
  isTotalLiability?: boolean
  currentPeriod?: number
  previousPeriod?: number
  change?: number
  percentChange?: number
}

export interface RowVariantProps {
  isHeader?: boolean
  isTotal?: boolean
  isTotalLiability?: boolean
  isTotalEquity?: boolean
  isAccount?: boolean
  item?: FinancialData
  isPositive: boolean
  isNegative: boolean
}